+++ packages/google_mobile_ads/lib/src/ad_containers.dart
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 import 'dart:async';
-import 'dart:io' show Platform;
 import 'dart:math';
 
 import 'package:flutter/foundation.dart';
@@ -95,7 +94,6 @@ class AdapterResponseInfo {
     required this.adapterClassName,
     required this.latencyMillis,
     required this.description,
-    required this.credentials,
     required this.adUnitMapping,
     this.adError,
   });
@@ -110,10 +108,6 @@ class AdapterResponseInfo {
 
   /// A log friendly string version of this object.
   final String description;
-
-  /// A string description of adapter credentials specified in the AdMob or Ad Manager UI.
-  @Deprecated('Use [adUnitMapping] instead')
-  final String credentials;
 
   /// Network configuration set on the AdMob UI.
   final Map<String, String> adUnitMapping;
@@ -148,40 +142,6 @@ class LoadAdError extends AdError {
   }
 }
 
-/// Location parameters that can be configured in an ad request.
-class LocationParams {
-  /// Location parameters that can be configured in an ad request.
-  const LocationParams({
-    required this.accuracy,
-    required this.longitude,
-    required this.latitude,
-    this.time,
-  });
-
-  /// The accuracy in meters.
-  final double accuracy;
-
-  /// The longitude in degrees.
-  final double longitude;
-
-  /// The latitude in degrees.
-  final double latitude;
-
-  /// The UTC time, in milliseconds since epoch (January 1, 1970).
-  ///
-  /// This is required on Android, and ignored on iOS.
-  final int? time;
-
-  @override
-  bool operator ==(Object other) {
-    return other is LocationParams &&
-        accuracy == other.accuracy &&
-        longitude == other.longitude &&
-        latitude == other.latitude &&
-        time == other.time;
-  }
-}
-
 /// Targeting info per the AdMob API.
 ///
 /// This class's properties mirror the native AdRequest API. See for example:
@@ -194,7 +154,6 @@ class AdRequest {
     this.neighboringContentUrls,
     this.nonPersonalizedAds,
     this.httpTimeoutMillis,
-    this.location,
     this.mediationExtrasIdentifier,
     this.extras,
   });
@@ -220,12 +179,6 @@ class AdRequest {
   ///
   /// This is only supported in Android. This value is ignored on iOS.
   final int? httpTimeoutMillis;
-
-  /// Location data.
-  ///
-  /// Used for mediation targeting purposes.
-  @Deprecated('Location is not used and will be deleted in a future release')
-  final LocationParams? location;
 
   /// String identifier used in providing mediation extras.
   ///
@@ -263,7 +216,6 @@ class AdManagerAdRequest extends AdRequest {
     bool? nonPersonalizedAds,
     int? httpTimeoutMillis,
     this.publisherProvidedId,
-    LocationParams? location,
     String? mediationExtrasIdentifier,
     Map<String, String>? extras,
   }) : super(
@@ -272,7 +224,6 @@ class AdManagerAdRequest extends AdRequest {
           neighboringContentUrls: neighboringContentUrls,
           nonPersonalizedAds: nonPersonalizedAds,
           httpTimeoutMillis: httpTimeoutMillis,
-          location: location,
           mediationExtrasIdentifier: mediationExtrasIdentifier,
           extras: extras,
         );
@@ -879,15 +830,6 @@ class BannerAd extends AdWithView {
   @override
   final BannerAdListener listener;
 
-  /// Check out developer pages for [Admob](https://developers.google.com/admob/flutter/test-ads)
-  /// and [AdManager](https://developers.google.com/ad-manager/mobile-ads-sdk/flutter/test-ads) for
-  /// demo ad units that point to specific test creatives for each format.
-  @Deprecated(
-      'Use test ad unit ids from the developer page while creating the ad.')
-  static final String testAdUnitId = Platform.isAndroid
-      ? 'ca-app-pub-3940256099942544/6300978111'
-      : 'ca-app-pub-3940256099942544/2934735716';
-
   @override
   Future<void> load() async {
     await instanceManager.loadBannerAd(this);
@@ -1056,15 +998,6 @@ class NativeAd extends AdWithView {
   /// Options to configure the native ad request.
   final NativeAdOptions? nativeAdOptions;
 
-  /// Check out developer pages for [Admob](https://developers.google.com/admob/flutter/test-ads)
-  /// and [AdManager](https://developers.google.com/ad-manager/mobile-ads-sdk/flutter/test-ads) for
-  /// demo ad units that point to specific test creatives for each format.
-  @Deprecated(
-      'Use test ad unit ids from the developer page while creating the ad.')
-  static final String testAdUnitId = Platform.isAndroid
-      ? 'ca-app-pub-3940256099942544/2247696110'
-      : 'ca-app-pub-3940256099942544/3986624511';
-
   @override
   Future<void> load() async {
     await instanceManager.loadNativeAd(this);
@@ -1091,15 +1024,6 @@ class InterstitialAd extends AdWithoutView {
 
   /// Callbacks to be invoked when ads show and dismiss full screen content.
   FullScreenContentCallback<InterstitialAd>? fullScreenContentCallback;
-
-  /// Check out developer pages for [Admob](https://developers.google.com/admob/flutter/test-ads)
-  /// and [AdManager](https://developers.google.com/ad-manager/mobile-ads-sdk/flutter/test-ads) for
-  /// demo ad units that point to specific test creatives for each format.
-  @Deprecated(
-      'Use test ad unit ids from the developer page while creating the ad.')
-  static final String testAdUnitId = Platform.isAndroid
-      ? 'ca-app-pub-3940256099942544/1033173712'
-      : 'ca-app-pub-3940256099942544/4411468910';
 
   /// Loads an [InterstitialAd] with the given [adUnitId] and [request].
   static Future<void> load({
@@ -1180,7 +1104,6 @@ class RewardedAd extends AdWithoutView {
     required String adUnitId,
     required this.rewardedAdLoadCallback,
     required this.request,
-    this.serverSideVerificationOptions,
   })  : adManagerRequest = null,
         super(adUnitId: adUnitId);
 
@@ -1191,7 +1114,6 @@ class RewardedAd extends AdWithoutView {
     required String adUnitId,
     required this.rewardedAdLoadCallback,
     required this.adManagerRequest,
-    this.serverSideVerificationOptions,
   })  : request = null,
         super(adUnitId: adUnitId);
 
@@ -1204,18 +1126,6 @@ class RewardedAd extends AdWithoutView {
   /// Callbacks for events that occur when attempting to load an ad.
   final RewardedAdLoadCallback rewardedAdLoadCallback;
 
-  /// Check out developer pages for [Admob](https://developers.google.com/admob/flutter/test-ads)
-  /// and [AdManager](https://developers.google.com/ad-manager/mobile-ads-sdk/flutter/test-ads) for
-  /// demo ad units that point to specific test creatives for each format.
-  @Deprecated(
-      'Use test ad unit ids from the developer page while creating the ad.')
-  static final String testAdUnitId = Platform.isAndroid
-      ? 'ca-app-pub-3940256099942544/5224354917'
-      : 'ca-app-pub-3940256099942544/1712485313';
-
-  /// Optional [ServerSideVerificationOptions].
-  ServerSideVerificationOptions? serverSideVerificationOptions;
-
   /// Callbacks to be invoked when ads show and dismiss full screen content.
   FullScreenContentCallback<RewardedAd>? fullScreenContentCallback;
 
@@ -1227,13 +1137,11 @@ class RewardedAd extends AdWithoutView {
     required String adUnitId,
     required AdRequest request,
     required RewardedAdLoadCallback rewardedAdLoadCallback,
-    ServerSideVerificationOptions? serverSideVerificationOptions,
   }) async {
     RewardedAd rewardedAd = RewardedAd._(
         adUnitId: adUnitId,
         request: request,
-        rewardedAdLoadCallback: rewardedAdLoadCallback,
-        serverSideVerificationOptions: serverSideVerificationOptions);
+        rewardedAdLoadCallback: rewardedAdLoadCallback);
 
     await instanceManager.loadRewardedAd(rewardedAd);
   }
@@ -1243,13 +1151,11 @@ class RewardedAd extends AdWithoutView {
     required String adUnitId,
     required AdManagerAdRequest adManagerRequest,
     required RewardedAdLoadCallback rewardedAdLoadCallback,
-    ServerSideVerificationOptions? serverSideVerificationOptions,
   }) async {
     RewardedAd rewardedAd = RewardedAd._fromAdManagerRequest(
         adUnitId: adUnitId,
         adManagerRequest: adManagerRequest,
-        rewardedAdLoadCallback: rewardedAdLoadCallback,
-        serverSideVerificationOptions: serverSideVerificationOptions);
+        rewardedAdLoadCallback: rewardedAdLoadCallback);
 
     await instanceManager.loadRewardedAd(rewardedAd);
   }
@@ -1262,6 +1168,11 @@ class RewardedAd extends AdWithoutView {
   Future<void> show({required OnUserEarnedRewardCallback onUserEarnedReward}) {
     onUserEarnedRewardCallback = onUserEarnedReward;
     return instanceManager.showAdWithoutView(this);
+  }
+
+  /// Set [ServerSideVerificationOptions] for the ad.
+  Future<void> setServerSideOptions(ServerSideVerificationOptions options) {
+    return instanceManager.setServerSideVerificationOptions(options, this);
   }
 }
 
@@ -1282,7 +1193,6 @@ class RewardedInterstitialAd extends AdWithoutView {
     required String adUnitId,
     required this.rewardedInterstitialAdLoadCallback,
     required this.request,
-    this.serverSideVerificationOptions,
   })  : adManagerRequest = null,
         super(adUnitId: adUnitId);
 
@@ -1293,7 +1203,6 @@ class RewardedInterstitialAd extends AdWithoutView {
     required String adUnitId,
     required this.rewardedInterstitialAdLoadCallback,
     required this.adManagerRequest,
-    this.serverSideVerificationOptions,
   })  : request = null,
         super(adUnitId: adUnitId);
 
@@ -1305,9 +1214,6 @@ class RewardedInterstitialAd extends AdWithoutView {
 
   /// Callbacks for events that occur when attempting to load an ad.
   final RewardedInterstitialAdLoadCallback rewardedInterstitialAdLoadCallback;
-
-  /// Optional [ServerSideVerificationOptions].
-  ServerSideVerificationOptions? serverSideVerificationOptions;
 
   /// Callbacks to be invoked when ads show and dismiss full screen content.
   FullScreenContentCallback<RewardedInterstitialAd>? fullScreenContentCallback;
@@ -1321,13 +1227,11 @@ class RewardedInterstitialAd extends AdWithoutView {
     required AdRequest request,
     required RewardedInterstitialAdLoadCallback
         rewardedInterstitialAdLoadCallback,
-    ServerSideVerificationOptions? serverSideVerificationOptions,
   }) async {
     RewardedInterstitialAd rewardedInterstitialAd = RewardedInterstitialAd._(
         adUnitId: adUnitId,
         request: request,
-        rewardedInterstitialAdLoadCallback: rewardedInterstitialAdLoadCallback,
-        serverSideVerificationOptions: serverSideVerificationOptions);
+        rewardedInterstitialAdLoadCallback: rewardedInterstitialAdLoadCallback);
 
     await instanceManager.loadRewardedInterstitialAd(rewardedInterstitialAd);
   }
@@ -1338,15 +1242,13 @@ class RewardedInterstitialAd extends AdWithoutView {
     required AdManagerAdRequest adManagerRequest,
     required RewardedInterstitialAdLoadCallback
         rewardedInterstitialAdLoadCallback,
-    ServerSideVerificationOptions? serverSideVerificationOptions,
   }) async {
     RewardedInterstitialAd rewardedInterstitialAd =
         RewardedInterstitialAd._fromAdManagerRequest(
             adUnitId: adUnitId,
             adManagerRequest: adManagerRequest,
             rewardedInterstitialAdLoadCallback:
-                rewardedInterstitialAdLoadCallback,
-            serverSideVerificationOptions: serverSideVerificationOptions);
+                rewardedInterstitialAdLoadCallback);
 
     await instanceManager.loadRewardedInterstitialAd(rewardedInterstitialAd);
   }
@@ -1359,6 +1261,11 @@ class RewardedInterstitialAd extends AdWithoutView {
   Future<void> show({required OnUserEarnedRewardCallback onUserEarnedReward}) {
     onUserEarnedRewardCallback = onUserEarnedReward;
     return instanceManager.showAdWithoutView(this);
+  }
+
+  /// Set [ServerSideVerificationOptions] for the ad.
+  Future<void> setServerSideOptions(ServerSideVerificationOptions options) {
+    return instanceManager.setServerSideVerificationOptions(options, this);
   }
 }